+++ src/Ext/Building/Body.h
@@ -39,6 +39,7 @@ public:
 		std::optional<int> CurrentLaserWeaponIndex;
 		int PoweredUpToLevel; // Distinct from UpgradeLevel, and set to highest PowersUpToLevel out of applied upgrades regardless of how many are currently applied to this building.
 		SuperClass* EMPulseSW;
+		CellStruct RandomEMPTarget;
 
 		ExtData(BuildingClass* OwnerObject) : Extension<BuildingClass>(OwnerObject)
 			, TypeExtData { nullptr }
@@ -54,6 +55,7 @@ public:
 			, CurrentLaserWeaponIndex {}
 			, PoweredUpToLevel { 0 }
 			, EMPulseSW {}
+			, RandomEMPTarget { CellStruct::Empty }
 		{ }
 
 		void DisplayIncomeString();