+++ docs/Whats-New.md
@@ -453,6 +453,7 @@ New:
 - Battle Points economy for super weapons (by FS-21)
 - New EVA voice after deploying a building (by FS-21)
 - Engineer logics on Warheads (by FS-21)
+- `EMPulseCannon.InaccurateRadius` for projectiles in `EMPulseCannon=yes` superweapons and `Ammo` support (by FS-21)
 
 Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)